Encapsulated some components.
